xclaw/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xclaw"
version = "0.4.0"
description = "X (Twitter) CLI for AI agents — OAuth 1.0a + free scout"
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "requests>=2.28",
    "requests-oauthlib>=1.3",
    "python-dotenv>=1.0",
    "twikit>=2.0",
]

[project.scripts]
xclaw = "xclaw.cli:cli"

[tool.setuptools.packages.find]
include = ["xclaw*"]
